feat: add service worker to suppoer offline page
Add service worker to fetch offline page (i.e. /banking-web/offline)
once page is loaded

To support that, a route (i.e. /offline) is added to serve html to
the cache; The route is protected with logic not to let non service-worker
requests access the page by checking header's referer